# api/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "API for managing sandboxes, organizations, billing, and hosts"
  title: Fluid API
  version: "1.0"
servers:
- url: //localhost:8081/v1
paths:
...
  /orgs/{slug}/hosts:
    get:
      description: List all connected sandbox hosts
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List hosts
      tags:
      - Hosts
  /orgs/{slug}/hosts/{hostID}:
    get:
      description: Get details of a specific connected host
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Host ID
        in: path
        name: hostID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orchestrator.HostInfo"
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
      security:
      - CookieAuth: []
      summary: Get host
      tags:
      - Hosts
  /orgs/{slug}/hosts/tokens:
    get:
      description: List all host tokens for the organization
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List host tokens
      tags:
      - Host Tokens
    post:
      description: Generate a new host authentication token (owner or admin only)
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.createHostTokenRequest"
        description: Token details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.hostTokenResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Create host token
      tags:
      - Host Tokens
      x-codegen-request-body-name: request
  /orgs/{slug}/hosts/tokens/{tokenID}:
    delete:
      description: Delete a host token (owner or admin only)
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Token ID
        in: path
        name: tokenID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
      security:
      - CookieAuth: []
      summary: Delete host token
      tags:
      - Host Tokens
...
components:
  schemas:
...
    orchestrator.HostInfo:
      example:
        hostname: hostname
        available_disk_mb: 1
        available_memory_mb: 5
        active_sandboxes: 0
        base_images:
        - base_images
        - base_images
        last_heartbeat: last_heartbeat
        available_cpus: 6
        host_id: host_id
        status: status
      properties:
        active_sandboxes:
          type: integer
        available_cpus:
          type: integer
        available_disk_mb:
          type: integer
        available_memory_mb:
          type: integer
        base_images:
          items:
            type: string
          type: array
        host_id:
          type: string
        hostname:
          type: string
        last_heartbeat:
          type: string
        status:
          type: string
      type: object
...
    rest.createHostTokenRequest:
      properties:
        name:
          type: string
      type: object
...
    rest.hostTokenResponse:
      example:
        name: name
        created_at: created_at
        id: id
        token: token
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        token:
          description: Only set on creation.
          type: string
      type: object
...
    rest.swaggerError:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
...